+++ packages/middlewares/src/index.js
@@ -8,4 +8,5 @@ export { default as serveStatic } from './serveStatic';
 export { default as passportAuth } from './passportAuth';
 export { default as passportCheck } from './passportCheck';
 export { default as userAgent } from './userAgent';
+export { default as localeDetect } from './localeDetect';
 export { default as views } from './views';